Maintained state management with zustand + added database dump
