+++ database/sudoku.sql
@@ -1,0 +1,77 @@
+-- phpMyAdmin SQL Dump
+-- version 5.2.1
+-- https://www.phpmyadmin.net/
+--
+-- Хост: 127.0.0.1:3307
+-- Время создания: Окт 31 2025 г., 11:08
+-- Версия сервера: 9.1.0
+-- Версия PHP: 8.3.14
+
+SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
+START TRANSACTION;
+SET time_zone = "+00:00";
+
+
+/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
+/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
+/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
+/*!40101 SET NAMES utf8mb4 */;
+
+--
+-- База данных: `sudoku`
+--
+
+-- --------------------------------------------------------
+
+--
+-- Структура таблицы `game_stats`
+--
+
+DROP TABLE IF EXISTS `game_stats`;
+CREATE TABLE IF NOT EXISTS `game_stats` (
+  `id` int NOT NULL AUTO_INCREMENT,
+  `user_id` int NOT NULL,
+  `difficulty` varchar(10) NOT NULL,
+  `time_spent` int NOT NULL,
+  `correct_cells` int NOT NULL,
+  `total_cells` int NOT NULL DEFAULT '81',
+  `score_percent` int NOT NULL,
+  `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
+  PRIMARY KEY (`id`),
+  KEY `user_id_idx` (`user_id`)
+) ENGINE=MyISAM AUTO_INCREMENT=15 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
+
+--
+-- Дамп данных таблицы `game_stats`
+--
+
+INSERT INTO `game_stats` (`id`, `user_id`, `difficulty`, `time_spent`, `correct_cells`, `total_cells`, `score_percent`, `created_at`) VALUES
+(4, 1, 'medium', 1, 35, 81, 43, '2025-10-29 10:49:05');
+
+-- --------------------------------------------------------
+
+--
+-- Структура таблицы `users`
+--
+
+DROP TABLE IF EXISTS `users`;
+CREATE TABLE IF NOT EXISTS `users` (
+  `id` int NOT NULL AUTO_INCREMENT,
+  `name` varchar(255) NOT NULL,
+  `email` varchar(255) NOT NULL,
+  `passwordHash` varchar(255) NOT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `email` (`email`)
+) ENGINE=InnoDB AUTO_INCREMENT=4 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
+
+--
+-- Дамп данных таблицы `users`
+--
+
+INSERT INTO `users` (`id`, `name`, `email`, `passwordHash`) VALUES
+(3, '12', '1@gmail.com', '$2y$10$Fa70UPXJI41kjTcbriEDFOQpp8CCVRwyEt8vYdpjC4ZnZAwfZV3ue');
+COMMIT;
+
+/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
+/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
+/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
